merge Nam(nam + backup1)

+++ app/controllers/admin/Nganh.php
@@ -14,6 +14,11 @@ class Nganh extends Controller {
     }
 
     function list_nganh() {
+
+        checkPermission(['Quản lý']);
+
+
+
         $search = isGet() ? (!empty($_GET['search']) ? $_GET['search'] : '') : '';
         $result = $this->model->getAll($search);
     
@@ -28,6 +33,11 @@ class Nganh extends Controller {
     }
 
     public function add_nganh() {
+
+
+        checkPermission(['Quản lý']);
+
+
         if (isPost()) {
             $filteredPost = filter();
             if (!$this->model->isDuplicateNganhId($filteredPost['major_id'])) {
@@ -43,6 +53,11 @@ class Nganh extends Controller {
     }
 
     public function edit_nganh($id = '') {
+
+
+        checkPermission(['Quản lý']);
+
+
         if (isPost()) {
             $filteredPost = filter(); 
             $this->model->updateNganh($id, $filteredPost); 
@@ -64,6 +79,9 @@ class Nganh extends Controller {
     
 
     public function delete_nganh($id = '') {
+
+        checkPermission(['Quản lý']);
+
         if ($this->model->deleteNganh($id)) {
             echo "<script>alert('Xóa thành công')</script>";
         } else {
